+++ extension/src/panel/routes/ConnectButton/injectedWallet/index.ts
@@ -1,0 +1,1 @@
+export { InjectedWalletConnect } from './InjectedWalletConnect'
