+++ extension/src/panel/routes/ConnectButton/walletConnect/index.ts
@@ -1,0 +1,1 @@
+export { WalletConnectConnect } from './WalletConnectConnect'
