+++ extension/src/panel/routes/ConnectButton/walletConnect/wallet-connect-logo.png
